[server] coap block2 GET response

+++ core/er-coap-13/er-coap-13.h
@@ -258,7 +258,7 @@ typedef struct {
   multi_option_t *uri_query;
   uint8_t if_none_match;
 
-  uint16_t payload_len;
+  uint32_t payload_len;
   uint8_t *payload;
 
 } coap_packet_t;
